refactor: Error boundary component
