+++ src/components/ErrorBoundary/types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react';
+
+interface IErrorBoundaryProps {
+  children: ReactNode;
+}
+
+interface IErrorBoundaryState {
+  hasError: boolean;
+}
+
+export type { IErrorBoundaryProps, IErrorBoundaryState };
